backend/requirements.txt
# --- Cấu hình kho tải (Giữ nguyên để tải bản CPU nhẹ) ---
--extra-index-url https://download.pytorch.org/whl/cpu

# --- Framework Web & API ---
# Nâng cấp lên bản mới ổn định để vá lỗi bảo mật
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.4

# --- Database (Đã xóa SQLAlchemy/Alembic/Psycopg2) ---
# Prisma Client Python (Nên khớp version với prisma cli ở package.json nếu có)
prisma==0.13.1

# --- AI & Xử lý ảnh ---
# PyTorch 2.2.x ổn định hơn và nhanh hơn 2.1 trên CPU
torch==2.2.2+cpu
torchvision==0.17.2+cpu
# Facenet dùng để trích xuất đặc trưng khuôn mặt
facenet-pytorch==2.5.3
numpy==1.26.4
pillow==10.3.0
# Bản headless là chuẩn cho Docker (không cần GUI)
opencv-python-headless==4.10.0.82

# --- Vector Database & Data ---
# Dùng faiss-cpu để tìm kiếm khuôn mặt nhanh
faiss-cpu==1.8.0
pandas==2.2.2

packaging

# --- Tiện ích ---
python-dotenv==1.0.1
requests==2.32.3
email-validator==2.1.1

# --- Security ---
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0